refactor: altera o nome do diretório 'database' para 'config'
